move check functions to tests, don't use addresses from config file
