+++ tests/_test_fallback_rollback_l1.py
@@ -1,7 +1,8 @@
 """
 Tests for voting xx/xx/2024
 """
-from scripts.upgrade_2024_10_08 import start_vote, check_pre_upgrade_state, check_post_upgrade_state
+from scripts.upgrade_2024_10_08 import start_vote
+from test_2024_10_08 import check_pre_upgrade_state, check_post_upgrade_state
 from scripts.fallback_rollback_l1 import start_vote as start_vote_fallback
 from brownie import interface
 from utils.test.tx_tracing_helpers import *
